refactor: enum test각 도메인으로 분리
